feat: extend JwtPolicy for Logout TTL calculation and Add Auth Exceptions

+++ auth/src/main/java/com/lxp/auth/application/local/port/required/usecase/LogoutUserUseCase.java
@@ -3,6 +3,6 @@ package com.lxp.auth.application.local.port.required.usecase;
 @FunctionalInterface
 public interface LogoutUserUseCase {
 
-    void logout(String accessToken, int remainingSeconds);
+    void logout(String accessToken, long remainingSeconds);
 
 }
